+++ app/api/ada/lib/storage/database/walletTypes/bip44/api/utils.js
@@ -1,33 +1,40 @@
 // @flow
 
-import * as Bip44Tables from '../tables';
-import type { Bip44AddressInsert, } from '../tables';
+import {
+  Bip44RootSchema,
+  Bip44PurposeSchema,
+  Bip44CoinTypeSchema,
+  Bip44AccountSchema,
+  Bip44ChainSchema,
+  Bip44AddressSchema,
+} from '../../common/tables';
+import type { Bip44AddressInsert, } from '../../common/tables';
 import type { TreeInsert, } from './write';
 
 export const Bip44DerivationLevels = Object.freeze({
   ROOT: {
     level: 0,
-    table: Bip44Tables.Bip44RootSchema,
+    table: Bip44RootSchema,
   },
   PURPOSE: {
     level: 1,
-    table: Bip44Tables.Bip44PurposeSchema,
+    table: Bip44PurposeSchema,
   },
   COIN_TYPE: {
     level: 2,
-    table: Bip44Tables.Bip44CoinTypeSchema,
+    table: Bip44CoinTypeSchema,
   },
   ACCOUNT: {
     level: 3,
-    table: Bip44Tables.Bip44AccountSchema,
+    table: Bip44AccountSchema,
   },
   CHAIN: {
     level: 4,
-    table: Bip44Tables.Bip44ChainSchema,
+    table: Bip44ChainSchema,
   },
   ADDRESS: {
     level: 5,
-    table: Bip44Tables.Bip44AddressSchema,
+    table: Bip44AddressSchema,
   },
 });
 export const Bip44TableMap = new Map<number, string>(
@@ -46,12 +53,12 @@ export const Bip44TableMap = new Map<number, string>(
  * Since we cannot statically determine which level will be used, we just lock all tables.
  */
 export const allBip44DerivationTables = {
-  [Bip44Tables.Bip44RootSchema.name]: Bip44Tables.Bip44RootSchema,
-  [Bip44Tables.Bip44PurposeSchema.name]: Bip44Tables.Bip44PurposeSchema,
-  [Bip44Tables.Bip44CoinTypeSchema.name]: Bip44Tables.Bip44CoinTypeSchema,
-  [Bip44Tables.Bip44AccountSchema.name]: Bip44Tables.Bip44AccountSchema,
-  [Bip44Tables.Bip44ChainSchema.name]: Bip44Tables.Bip44ChainSchema,
-  [Bip44Tables.Bip44AddressSchema.name]: Bip44Tables.Bip44AddressSchema,
+  [Bip44RootSchema.name]: Bip44RootSchema,
+  [Bip44PurposeSchema.name]: Bip44PurposeSchema,
+  [Bip44CoinTypeSchema.name]: Bip44CoinTypeSchema,
+  [Bip44AccountSchema.name]: Bip44AccountSchema,
+  [Bip44ChainSchema.name]: Bip44ChainSchema,
+  [Bip44AddressSchema.name]: Bip44AddressSchema,
 };
 
 export function flattenInsertTree(
