+++ app/api/ada/lib/storage/database/walletTypes/cip1852/tables.js
@@ -2,6 +2,7 @@
 
 import type { lf$schema$Builder } from 'lovefield';
 
+
 export const populateCip1852Db = (schemaBuilder: lf$schema$Builder) => {
   // TODO
 };
